# backend/swagger.yaml
openapi: 3.0.0
info:
  title: CoffeeMaker
  description: Node.js, mongoDB CoffeeMaker with login/registration system
  version: 1.0.0
servers:
  - url: http://localhost:5050
    description: Local development server
...
paths:
...
  /api/votes:
    get:
      tags: 
      - Votes
      summary: List all votes
      description: User & Admin
      operationId: getVotes
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    [
                        {
                            "_id": "6288e0a6b3076a49e337fe3b",
                            "brewId": "6288b3fddde43e0c28357b2d",
                            "username": "myusername",
                            "value": 3,
                            "date": "2022-05-21T12:52:54.134Z",
                            "__v": 0
                        }
                    ]
        "400":
          description: Not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    post:
      tags: 
      - Votes
      summary: Vote for brew
      description: User & Admin
      operationId: postVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brewId:
                  type: string
                  example: 6288b3fddde43e0c28357b2d
                username:
                  type: string
                  example: myusername
                value:
                  type: number
                  example: 3
              required:
                - brewId
                - username
                - value
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "brewId": "6288b3fddde43e0c28357b2d",
                        "username": "myusername",
                        "value": 3,
                        "_id": "6288e0a6b3076a49e337fe3b",
                        "date": "2022-05-21T12:52:54.134Z",
                        "__v": 0
                    }
        "400":
          description: Please add all fields
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: User has already voted for this brew
        "500":
          description: Internal server error
    put:
      tags: 
      - Votes
      summary: Update specific vote
      description: User & Admin
      operationId: updateVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brewId:
                  type: string
                  example: 6288b3fddde43e0c28357b2d
                username:
                  type: string
                  example: myusername
                value:
                  type: number
                  example: 5
              required:
                - brewId
                - username
                - value
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                example:
                  value: |-
                    {
                        "_id": "6288e0a6b3076a49e337fe3b",
                        "brewId": "6288b3fddde43e0c28357b2d",
                        "username": "myusername",
                        "value": 5,
                        "date": "2022-05-21T12:52:54.134Z",
                        "__v": 0
                    }
        "400":
          description: Not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
  /api/votes/my-votes:
    get:
      tags: 
      - Votes
      summary: Get all votes from user
      description: User & Admin
      operationId: getMyVotes
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    [
                        {
                            "_id": "6288e0a6b3076a49e337fe3b",
                            "brewId": "6288b3fddde43e0c28357b2d",
                            "username": "myusername",
                            "value": 3,
                            "date": "2022-05-21T12:52:54.134Z",
                            "__v": 0
                        },
                        {
                            "_id": "6288e8ade8c870223d4c8db5",
                            "brewId": "6288b3fddde43e0c28357b2e",
                            "username": "myusername",
                            "value": 3,
                            "date": "2022-05-21T13:27:09.215Z",
                            "__v": 0
                        }
                    ]
        "400":
          description: Votes not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
  /api/votes/brew-specific:
    get:
      tags: 
      - Votes
      summary: Get all votes for brew
      description: User & Admin
      operationId: getBrewVotes
      parameters:
        - in: query
          name: brewId
          schema:
            type: integer
          description: id of brew
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    [
                        {
                            "_id": "6288e0a6b3076a49e337fe3b",
                            "brewId": "6288b3fddde43e0c28357b2d",
                            "username": "myusername",
                            "value": 3,
                            "date": "2022-05-21T12:52:54.134Z",
                            "__v": 0
                        },
                        {
                            "_id": "62892211f3930f516a808c1b",
                            "brewId": "6288b3fddde43e0c28357b2d",
                            "username": "anotherusername",
                            "value": 4,
                            "date": "2022-05-21T17:32:01.872Z",
                            "__v": 0
                        },
                        {
                            "_id": "62892d624c4588792bf968ad",
                            "brewId": "6288b3fddde43e0c28357b2d",
                            "username": "anotherusername6",
                            "value": 4,
                            "date": "2022-05-21T18:20:18.459Z",
                            "__v": 0
                        }
                    ]
        "400":
          description: Votes not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: >-
        Access token is missing or invalid, or the user does not have access to
        perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not Found